Remove isAPIErrorResponse and SafeParseJSON exports from types
