+++ front/components/assistant/conversation/ConversationError.tsx
@@ -8,7 +8,8 @@ import {
 import type { ComponentType } from "react";
 
 import type { ConversationError } from "@app/types";
-import { isAPIErrorResponse, safeParseJSON } from "@app/types";
+import { isAPIErrorResponse } from "@app/types";
+import { safeParseJSON } from "@app/types/shared/utils/json_utils";
 
 interface ConversationErrorProps {
   error: ConversationError;
