+++ front/lib/api/llm/clients/mistral/utils/conversation_to_mistral.ts
@@ -13,13 +13,13 @@ import type {
   Content,
   ModelMessageTypeMultiActionsWithoutContentFragment,
 } from "@app/types";
-import { safeParseJSON } from "@app/types";
+import { assertNever } from "@app/types/shared/utils/assert_never";
 import type {
   AgentFunctionCallContentType,
   AgentReasoningContentType,
   AgentTextContentType,
 } from "@app/types/assistant/agent_message_content";
-import { assertNever } from "@app/types/shared/utils/assert_never";
+import { safeParseJSON } from "@app/types/shared/utils/json_utils";
 
 export function sanitizeToolCallId(id: string): string {
   // Replace anything not a-zA-Z-0-9 with 0 as mistral enforces that but function_call_id can
