+++ front/lib/swr/swr.ts
@@ -16,7 +16,8 @@ import useSWRInfinite from "swr/infinite";
 
 import { COMMIT_HASH } from "@app/lib/commit-hash";
 import { clientFetch } from "@app/lib/egress/client";
-import { isAPIErrorResponse, safeParseJSON } from "@app/types";
+import { isAPIErrorResponse } from "@app/types/error";
+import { safeParseJSON } from "@app/types/shared/utils/json_utils";
 
 const EMPTY_ARRAY = Object.freeze([]);
