+++ front/types/shared/utils/string_utils.ts
@@ -2,8 +2,6 @@ import removeMarkdown from "remove-markdown";
 
 import { replaceContentNodeMarkdownWithQuotedTitle } from "@app/lib/content_nodes";
 import { replaceMentionsWithAt } from "@app/lib/mentions/format";
-import type { Result } from "@app/types/shared/result";
-import { Err, Ok } from "@app/types/shared/result";
 
 /**
  * Substring that ensures we don't cut a string in the middle of a unicode
@@ -100,20 +98,6 @@ export function truncate(text: string, length: number, omission = "...") {
   return text.length > length
     ? `${text.substring(0, length - omission.length)}${omission}`
     : text;
-}
-
-export function safeParseJSON(str: string): Result<object | null, Error> {
-  try {
-    const res = JSON.parse(str);
-
-    return new Ok(res);
-  } catch (err) {
-    if (err instanceof Error) {
-      return new Err(err);
-    }
-
-    return new Err(new Error("Unexpected error: JSON parsing failed."));
-  }
 }
 
 export function stripNullBytes(text: string): string {
